fix: Add generic function for updating the node for a wallet type, move ethereum transaction error fees handler to a new file
