+++ assets/ethereum_server_list.yml
@@ -8,3 +8,5 @@
   uri: eth-mainnet.public.blastapi.io
 -
   uri: eth.nownodes.io
+-
+  uri: ethereum.publicnode.com
