+++ lib/view_model/send/send_view_model.dart
@@ -1,5 +1,6 @@
 import 'package:cake_wallet/di.dart';
 import 'package:cake_wallet/entities/contact.dart';
+import 'package:cake_wallet/entities/evm_transaction_error_fees_handler.dart';
 import 'package:cake_wallet/entities/priority_for_wallet_type.dart';
 import 'package:cake_wallet/entities/transaction_description.dart';
 import 'package:cake_wallet/ethereum/ethereum.dart';
@@ -502,9 +503,7 @@ abstract class SendViewModelBase extends WalletChangeListenerViewModel with Stor
         nano!.updateTransactions(wallet);
       }
 
-
       if (pendingTransaction!.id.isNotEmpty) {
-
         final descriptionKey = '${pendingTransaction!.id}_${wallet.walletAddresses.primaryAddress}';
         _settingsStore.shouldSaveRecipientAddress
             ? await transactionDescriptionBox.add(TransactionDescription(
@@ -693,7 +692,8 @@ abstract class SendViewModelBase extends WalletChangeListenerViewModel with Stor
       }
 
       if (errorMessage.contains('insufficient funds')) {
-        final parsedErrorMessageResult = parseEthereumFeesErrorMessage(
+        final parsedErrorMessageResult =
+            EVMTransactionErrorFeesHandler.parseEthereumFeesErrorMessage(
           errorMessage,
           _fiatConversationStore.prices[currency]!,
         );
@@ -783,90 +783,4 @@ abstract class SendViewModelBase extends WalletChangeListenerViewModel with Stor
 
     return false;
   }
-
-  EVMTransactionErrorFeesHandler parseEthereumFeesErrorMessage(
-    String errorMessage,
-    double assetPriceUsd,
-  ) {
-    final EVMTransactionErrorFeesHandler result;
-
-    // Define Regular Expressions to extract the numerical values
-    RegExp balanceRegExp = RegExp(r'balance (\d+)');
-    RegExp txCostRegExp = RegExp(r'tx cost (\d+)');
-    RegExp overshotRegExp = RegExp(r'overshot (\d+)');
-
-    // Match the patterns in the error message
-    Match? balanceMatch = balanceRegExp.firstMatch(errorMessage);
-    Match? txCostMatch = txCostRegExp.firstMatch(errorMessage);
-    Match? overshotMatch = overshotRegExp.firstMatch(errorMessage);
-
-    // Check if all required values are found
-    if (balanceMatch != null && txCostMatch != null && overshotMatch != null) {
-      // Extract the numerical strings
-      String balanceStr = balanceMatch.group(1)!;
-      String txCostStr = txCostMatch.group(1)!;
-      String overshotStr = overshotMatch.group(1)!;
-
-      // Parse the numerical strings to BigInt
-      BigInt balanceWei = BigInt.parse(balanceStr);
-      BigInt txCostWei = BigInt.parse(txCostStr);
-      BigInt overshotWei = BigInt.parse(overshotStr);
-
-      // Convert wei to ETH (1 ETH = 1e18 wei)
-      double balanceEth = balanceWei.toDouble() / 1e18;
-      double txCostEth = txCostWei.toDouble() / 1e18;
-      double overshotEth = overshotWei.toDouble() / 1e18;
-
-      // Calculate the USD values
-      double balanceUsd = balanceEth * assetPriceUsd;
-      double txCostUsd = txCostEth * assetPriceUsd;
-      double overshotUsd = overshotEth * assetPriceUsd;
-
-      result = EVMTransactionErrorFeesHandler(
-        balanceWei: balanceWei.toString(),
-        balanceEth: balanceEth.toString().substring(0, 12),
-        balanceUsd: balanceUsd.toString().substring(0,4),
-        txCostWei: txCostWei.toString(),
-        txCostEth: txCostEth.toString().substring(0, 12),
-        txCostUsd: txCostUsd.toString().substring(0,4),
-        overshotWei: overshotWei.toString(),
-        overshotEth: overshotEth.toString().substring(0, 12),
-        overshotUsd: overshotUsd.toString().substring(0,4),
-      );
-    } else {
-      // If any value is missing, return an error message
-      result = EVMTransactionErrorFeesHandler(error: 'Could not parse the error message.');
-    }
-
-    return result;
-  }
-}
-
-class EVMTransactionErrorFeesHandler {
-  EVMTransactionErrorFeesHandler({
-    this.balanceWei,
-    this.balanceEth,
-    this.balanceUsd,
-    this.txCostWei,
-    this.txCostEth,
-    this.txCostUsd,
-    this.overshotWei,
-    this.overshotEth,
-    this.overshotUsd,
-    this.error,
-  });
-
-  String? balanceWei;
-  String? balanceEth;
-  String? balanceUsd;
-
-  String? txCostWei;
-  String? txCostEth;
-  String? txCostUsd;
-
-  String? overshotWei;
-  String? overshotEth;
-  String? overshotUsd;
-
-  String? error;
 }
